+++ src/components/HomepageFeatures/featureCard.module.css
@@ -1,0 +1,27 @@
+.card {
+  border-radius: 11px;
+  background-color: var(--ifm-card-background-color);
+  display: flex;
+  flex-direction: column;
+  transition: all 0.2s ease-in-out;
+  padding: 2.5rem 1.5rem;
+}
+
+[data-theme="dark"] .card {
+  color: var(--ifm-color-gray-500);
+}
+
+[data-theme="light"] .card {
+  color: #000000;
+}
+
+.card:hover {
+  text-decoration: none;
+  color: var(--ifm-color-gray-500);
+  transform: scale(1.05);
+}
+
+.heading{
+    display: flex;
+    justify-content: space-between;
+}
