+++ src/components/HomepageFeatures/featuresSection.module.css
@@ -1,0 +1,34 @@
+.features {
+  display: flex;
+  padding: 6rem 0 2rem 0;
+  width: 100%;
+}
+
+[data-theme="dark"] .featureSvg {
+  height: 200px;
+  width: 200px;
+  fill: #ffffff;
+}
+
+[data-theme="light"] .featureSvg {
+  height: 200px;
+  width: 200px;
+  fill: #000000;
+}
+
+.cards {
+  display: grid;
+  grid-template-columns: 1fr;
+  /* grid-template-rows: 1fr; */
+  grid-column-gap: 0px;
+  grid-row-gap: 1.5em;
+}
+
+@media (min-width: 996px) {
+  .cards {
+    grid-template-columns: repeat(3, 1fr);
+    grid-template-rows: 1fr;
+    grid-column-gap: 1.5em;
+    grid-row-gap: 0px;
+  }
+}
