+++ src/components/HomepageFeatures/hero.module.css
@@ -1,7 +1,6 @@
 .heroSection {
   background-color: var(--ifm-card-background-color);
   border-radius: 6px;
-  /* min-height: 400px; */
 }
 
 .content {
@@ -28,6 +27,12 @@
   text-align: center;
 }
 
+.heading {
+  font-size: 48px;
+  font-weight: var(--ifm-font-weight-semibold);
+  text-align: center;
+}
+
 @media (min-width: 996px) {
   .heroSection {
     background-image: url("/img/landing/splash.png");
@@ -41,11 +46,9 @@
   .description {
     text-align: left;
   }
-}
-
-.heading {
-  font-size: 48px;
-  font-weight: 500;
+  .heading {
+    text-align: left;
+  }
 }
 
 .button {
@@ -53,10 +56,12 @@
   color: #000000;
   padding: 10px 32px 10px 32px;
   font-size: 20px;
-  font-weight: 500;
+  font-weight: var(--ifm-font-weight-semibold);
+  transition: all .2s ease-in-out;
 }
 .button:hover {
   text-decoration: none;
+  transform: scale(1.1);
 }
 
 [data-theme="dark"] .button {
