+++ src/components/HomepageFeatures/heroSection.tsx
@@ -1,13 +1,16 @@
 import classes from "./hero.module.css";
 import clsx from "clsx";
 import Link from "@docusaurus/Link";
+import Heading from "@theme/Heading";
 
 export default function HeroSection() {
   return (
     <section className={classes.heroSection}>
       <div className={clsx(classes.content, "container")}>
         <div className={classes.callToAction}>
-          <h1 className={classes.heading}>Flare Developer Hub</h1>
+          <Heading as="h1" className={classes.heading}>
+            Flare Developer Hub
+          </Heading>
           <p className={classes.description}>
             The decentralization origin for Flare builders. Written by builders,
             for builders
