--- src/components/HomepageFeatures/styles.module.css
@@ -1,17 +1,0 @@
-.features {
-  display: flex;
-  padding: 2rem 0;
-  width: 100%;
-}
-
-[data-theme="dark"] .featureSvg {
-  height: 200px;
-  width: 200px;
-  fill: #ffffff;
-}
-
-[data-theme="light"] .featureSvg {
-  height: 200px;
-  width: 200px;
-  fill: #000000;
-}
