+++ src/pages/index.tsx
@@ -1,12 +1,12 @@
-import clsx from "clsx";
 import Link from "@docusaurus/Link";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import Layout from "@theme/Layout";
-import HomepageFeatures from "@site/src/components/HomepageFeatures";
 import Heading from "@theme/Heading";
+import Layout from "@theme/Layout";
+import clsx from "clsx";
 
-import styles from "./index.module.css";
+import FeaturesSection from "../components/HomepageFeatures/featuresSection";
 import HeroSection from "../components/HomepageFeatures/heroSection";
+import styles from "./index.module.css";
 
 function HomepageHeader() {
   const { siteConfig } = useDocusaurusContext();
@@ -35,7 +35,7 @@ export default function Home(): JSX.Element {
         <HeroSection />
         {/* <HomepageHeader /> */}
 
-        <HomepageFeatures />
+        <FeaturesSection />
       </main>
     </Layout>
   );
